공통 모듈 추가 및 기존 인증 기능을 인증 모듈로 이동
- 인증 모듈로 이동한 대상
    * 유저
    * 업체
    * 인증
