+++ growing-shop-auth/src/main/java/com/example/growingshopauth/auth/dto/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.example.growingshop.domain.auth.dto;
+package com.example.growingshopauth.dto;
 
 import lombok.*;
