+++ growing-shop-auth/src/main/java/com/example/growingshopauth/auth/dto/RoleRequest.java
@@ -1,4 +1,4 @@
-package com.example.growingshop.domain.auth.dto;
+package com.example.growingshopauth.dto;
 
 import com.example.growingshop.domain.auth.domain.HttpMethod;
 import com.example.growingshop.domain.auth.domain.Policies;
