+++ growing-shop-auth/src/main/java/com/example/growingshopauth/company/domain/CompanyStatus.java
