+++ growing-shop-auth/src/main/java/com/example/growingshopauth/config/error/GlobalExceptionHandler.java
@@ -1,29 +1,16 @@
-package com.example.growingshop.global.error.config;
+package com.example.growingshopauth.config.error;
 
-import com.example.growingshop.global.error.exception.InvalidJwtTokenException;
-import com.example.growingshop.global.error.exception.NotAllowPathException;
-import com.example.growingshop.global.error.exception.NotFoundUserException;
+import com.example.growingshopauth.config.error.exception.InvalidJwtTokenException;
+import com.example.growingshopauth.config.error.exception.NotAllowPathException;
+import com.example.growingshopauth.config.error.exception.NotFoundUserException;
+import com.example.growingshopcommon.config.error.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import javax.persistence.EntityNotFoundException;
-
 @RestControllerAdvice
 public class GlobalExceptionHandler {
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse illegalArgumentException(IllegalArgumentException exception) {
-        return new ErrorResponse(exception.getMessage());
-    }
-
-    @ExceptionHandler(EntityNotFoundException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse entityNotFoundException(EntityNotFoundException exception) {
-        return new ErrorResponse(exception.getMessage());
-    }
 
     @ExceptionHandler(NotFoundUserException.class)
     @ResponseStatus(HttpStatus.UNAUTHORIZED)
