+++ growing-shop-auth/src/main/java/com/example/growingshopauth/user/domain/UserType.java
