+++ growing-shop-common/build.gradle
@@ -1,0 +1,1 @@
+version = '0.0.1-SNAPSHOT'
