+++ growing-shop-common/src/main/java/com/example/growingshopcommon/GrowingShopCommonApplication.java
@@ -1,0 +1,13 @@
+package com.example.growingshopcommon;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GrowingShopCommonApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GrowingShopCommonApplication.class, args);
+    }
+
+}
