+++ growing-shop-common/src/main/resources/local/user_init.sql
