+++ growing-shop-common/src/test/java/com/example/growingshopcommon/GrowingShopCommonApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.growingshopcommon;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GrowingShopCommonApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
